+++ .flake8
@@ -2,3 +2,4 @@
 extend-ignore = E501,E203
 max-line-length = 88
 max-complexity = 15
+per-file-ignores = __init__.py:F401
