+++ landingai/pipeline/image_source.py
@@ -1,14 +1,21 @@
 """A module that provides a set of abstractions and APIs for reading images from different sources."""
 
+from datetime import datetime
 import glob
 from collections.abc import Iterator
 from pathlib import Path
 import tempfile
 import shutil
-from typing import Iterator as IteratorType
+import threading
+import time
+from typing import Any, Iterator as IteratorType, Optional
 from typing import List, Union
 
-from landingai.vision_pipeline import FrameSet
+import cv2
+import numpy as np
+from pydantic import BaseModel, PrivateAttr
+
+from landingai.pipeline.frameset import Frame, FrameSet
 from landingai.io import sample_images_from_video
 from landingai.storage.data_access import fetch_from_uri
 
@@ -90,7 +97,7 @@ class ImageFolder(ImageSourceBase):
     def __iter__(self) -> IteratorType[FrameSet]:
         for img_path in self._image_paths:
             meta = {"image_path": img_path}
-            yield FrameSet.from_image(img_path, metadata=meta)
+            yield FrameSet.from_image(str(img_path), metadata=meta)
 
     def __len__(self) -> int:
         return len(self._image_paths)
@@ -110,8 +117,10 @@ class VideoFile(ImageSourceBase):
 
     Example:
     ```python
-        img_src = VideoFile("sample_images/surfers.mp4", samples_per_second=1)
-        frs = FrameSet()
+        import landingai.pipeline as pl
+
+        img_src = pl.image_source.VideoFile("sample_images/surfers.mp4", samples_per_second=1)
+        frs = pl.FrameSet()
         for i,frame in enumerate(img_src):
             if i>=3: # Fetch only 3 frames
                 break
@@ -119,7 +128,7 @@ class VideoFile(ImageSourceBase):
                 frame.run_predict(predictor=surfer_model)
                 .overlay_predictions()
             )
-        print(frs.get_class_counts())
+        print(pl.postprocessing.get_class_counts(frs))
     ```
     """
 
@@ -145,3 +154,144 @@ class VideoFile(ImageSourceBase):
 
     def __del__(self) -> None:
         shutil.rmtree(self._local_cache_dir)
+
+
+# openCV's default VideoCapture cannot drop frames so if the CPU is overloaded the stream will start to lag behind realtime.
+# This class creates a treaded capture implementation that can stay up to date wit the stream and decodes frames only on demand
+
+
+class NetworkedCamera(BaseModel):
+    """The NetworkCamera class can connect to RTSP and other live video sources in order to grab frames. The main concern is to be able to consume frames at the source speed and drop them as needed to ensure the application allday gets the lastes frame"""
+
+    stream_url: str
+    motion_detection_threshold: int
+    capture_interval: Union[float, None] = None
+    previous_frame: Union[Frame, None] = None
+    _last_capture_time: datetime = PrivateAttr()
+    _cap: Any = PrivateAttr()  # cv2.VideoCapture
+    _t: Any = PrivateAttr()  # threading.Thread
+    _t_lock: Any = PrivateAttr()  # threading.Lock
+    _t_running: bool = PrivateAttr()
+
+    def __init__(
+        self,
+        stream_url: str,
+        motion_detection_threshold: int = 0,
+        capture_interval: Optional[float] = None,
+        fps: Optional[int] = None,
+    ) -> None:
+        """
+        Parameters
+        ----------
+        stream_url : url to video source
+        motion_detection_threshold : If set to zero then motion detections is disabled. Any other value (0-100) will make the camera drop all images that don't have significant changes
+        capture_interval : Number of seconds to wait in between frames. If set to None, the NetworkedCamera will acquire images as fast as the source permits.
+        fps: Capture speed in frames per second. If set to None, the NetworkedCamera will acquire images as fast as the source permits.
+        """
+        if fps is not None and capture_interval is not None:
+            raise ValueError(
+                "The fps and capture_interval arguments cannot be set at the same time"
+            )
+        elif fps is not None:
+            capture_interval = 1 / fps
+
+        if capture_interval is not None and capture_interval < 1 / 30:
+            raise ValueError(
+                "The resulting fps cannot be more than 30 frames per second"
+            )
+        cap = cv2.VideoCapture(stream_url)
+        if not cap.isOpened():
+            cap.release()
+            raise Exception(f"Could not open stream ({stream_url})")
+        cap.set(
+            cv2.CAP_PROP_BUFFERSIZE, 2
+        )  # Limit buffering to 2 frames in order to avoid backlog (i.e. lag)
+
+        super().__init__(
+            stream_url=stream_url,
+            motion_detection_threshold=motion_detection_threshold,
+            capture_interval=capture_interval,
+        )
+        self._last_capture_time = datetime.now()
+        self._cap = cap
+        self._t_lock = threading.Lock()
+        self._t = threading.Thread(target=self._reader)
+        self._t.daemon = True
+        self._t_running = False
+        self._t.start()
+
+    def __del__(self) -> None:
+        self._t_running = False
+        self._t.join(timeout=10)
+        self._cap.release()
+
+    # grab frames as soon as they are available
+    def _reader(self) -> None:
+        self._t_running = True
+        # inter_frame_interval = 1 / self._cap.get(cv2.CAP_PROP_FPS)  # Get the source's framerate (FPS = 1/X)
+        inter_frame_interval = (
+            1 / 30
+        )  # Some sources miss report framerate so we use a conservative number
+        while self._t_running:
+            with self._t_lock:
+                ret = self._cap.grab()  # non-blocking call
+                if not ret:
+                    raise Exception(f"Connection to camera broken ({self.stream_url})")
+            time.sleep(inter_frame_interval)  # Limit acquisition speed
+
+    # retrieve latest frame
+    def get_latest_frame(self) -> "FrameSet":
+        """Return the most up to date frame by dropping all by the latest frame. This function is blocking"""
+        if self.capture_interval is not None:
+            t = datetime.now()
+            delta = (t - self._last_capture_time).total_seconds()
+            if delta <= self.capture_interval:
+                time.sleep(self.capture_interval - delta)
+        with self._t_lock:
+            ret, frame = self._cap.retrieve()  # non-blocking call
+            if not ret:
+                raise Exception(f"Connection to camera broken ({self.stream_url})")
+        self._last_capture_time = datetime.now()
+        if self.motion_detection_threshold > 0:
+            if self._detect_motion(frame):
+                return FrameSet.from_array(frame)
+            else:
+                return FrameSet()  # Empty frame
+
+        return FrameSet.from_array(frame)
+
+    def _detect_motion(self, frame: np.ndarray) -> bool:  # TODO Needs test cases
+        """ """
+        # Prepare image; grayscale and blur
+        prepared_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+        prepared_frame = cv2.GaussianBlur(src=prepared_frame, ksize=(5, 5), sigmaX=0)
+
+        if self.previous_frame is None:
+            # Save the result for the next invocation
+            self.previous_frame = prepared_frame
+            return True  # First frame; there is no previous one yet
+
+        # calculate difference and update previous frame TODO: don't assume the processed image is cached
+        diff_frame = cv2.absdiff(src1=self.previous_frame, src2=prepared_frame)
+        # Only take different areas that are different enough (>20 / 255)
+        thresh_frame = cv2.threshold(
+            src=diff_frame, thresh=20, maxval=255, type=cv2.THRESH_BINARY
+        )[1]
+        change_percentage = (
+            100 * cv2.countNonZero(thresh_frame) / (frame.shape[0] * frame.shape[1])
+        )
+        # print(f"Image change {change_percentage:.2f}%")
+        if change_percentage > self.motion_detection_threshold:
+            self.previous_frame = prepared_frame
+            return True
+        return False
+
+    # Make the class iterable. TODO Needs test cases
+    def __iter__(self) -> Any:
+        return self
+
+    def __next__(self) -> "FrameSet":
+        return self.get_latest_frame()
+
+    class Config:
+        arbitrary_types_allowed = True
