+++ landingai/storage/data_access.py
@@ -62,7 +62,6 @@ def fetch_from_uri(uri: str, **kwargs) -> Path:  # type: ignore
     """
     # TODO support other URIs
     # snowflake://stage/filename  (credentials will be passed on kwargs)
-
     r = urlparse(uri)
     if r.scheme == "" or r.scheme == "file":
         # The file is already local
